Remove rules module
We don't need it now that we have pydantic rules models.
